refactor: modularize lsmcp.ts into smaller components
- Extract tool registry to handle LSP tool registration
- Extract adapter setup for registering built-in adapters
- Extract legacy support functions for backward compatibility
- Extract CLI help functions to separate module
- Extract LSP server runner for server lifecycle management
- Reduce lsmcp.ts from 736 lines to 325 lines (56% reduction)
